+++ taskqueue.go
@@ -4,11 +4,11 @@ import (
 	"github.com/pkg/errors"
 )
 
-//DefaultQueueSize indicates the default size of the TaskQueue
+// DefaultQueueSize indicates the default size of the TaskQueue
 const DefaultQueueSize = 10
 
-//TaskQueue represents a queue that holds tasks which
-//are in waiting for workers from the pool
+// TaskQueue represents a queue that holds tasks which
+// are in waiting for workers from the pool1
 type TaskQueue struct {
 	//queue that holds tasks
 	que []Task
@@ -17,19 +17,19 @@ type TaskQueue struct {
 	totalTasks int
 }
 
-//NewTaskQueue returns new TaskQueue with the default capacity
+// NewTaskQueue returns new TaskQueue with the default capacity
 func NewTaskQueue() *TaskQueue {
 	wtQue := TaskQueue{que: make([]Task, 0, DefaultQueueSize), totalTasks: 0}
 	return &wtQue
 }
 
-//EnqueueTask puts a new task in the TaskQueue
+// EnqueueTask puts a new task in the TaskQueue
 func (wq *TaskQueue) EnqueueTask(task Task) {
 	wq.que = append(wq.que, task)
 	wq.totalTasks++
 }
 
-//DequeueTask returns a task and removes it from the TaskQueue
+// DequeueTask returns a task and removes it from the TaskQueue
 func (wq *TaskQueue) DequeueTask() (Task, error) {
 	if wq.totalTasks > 0 {
 		task := wq.que[0]
@@ -40,7 +40,7 @@ func (wq *TaskQueue) DequeueTask() (Task, error) {
 	return Task{}, errors.New("Queue is Empty")
 }
 
-//IsNotEmpty returns whether the TaskQueue is empty or not
+// IsNotEmpty returns whether the TaskQueue is empty or not
 func (wq *TaskQueue) IsNotEmpty() bool {
 	return wq.totalTasks > 0
 }
