+++ worker.go
@@ -2,18 +2,18 @@ package gohive
 
 import "log"
 
-//Represents a worker which is responsible to execute a task,
-//handle panics if any and after completing the task, notifies
-//back to the pool to pull new task from the TaskQueue
+// Represents a worker which is responsible to execute a task,
+// handle panics if any and after completing the task, notifies
+// back to the pool1 to pull new task from the TaskQueue
 type worker struct {
 	//channel that receives a task
 	taskChan chan func()
 
 	//reference back to the pool who owns this worker
-	pool *pool
+	pool *pool1
 }
 
-//This method executes the task
+// This method executes the task
 func (w *worker) run() {
 	go func() {
 		defer func() {
